+++ packages/editor/editor-mobile-bridge/src/editor/web-to-native/markState.ts
@@ -1,44 +1,46 @@
+import { TextFormattingState } from '@atlaskit/editor-core';
+
 export class MarkState {
   name: string;
   active: boolean;
   enabled: boolean;
 }
 
-export function valueOf(state): MarkState[] {
+export function valueOf(state: TextFormattingState): MarkState[] {
   let states: MarkState[] = [
     {
       name: 'strong',
-      active: state.strongActive,
+      active: state.strongActive || false,
       enabled: !state.strongDisabled,
     },
     {
       name: 'em',
-      active: state.emActive,
+      active: state.emActive || false,
       enabled: !state.emDisabled,
     },
     {
       name: 'code',
-      active: state.codeActive,
+      active: state.codeActive || false,
       enabled: !state.codeDisabled,
     },
     {
       name: 'underline',
-      active: state.underlineActive,
+      active: state.underlineActive || false,
       enabled: !state.underlineDisabled,
     },
     {
       name: 'strike',
-      active: state.strikeActive,
+      active: state.strikeActive || false,
       enabled: !state.strongDisabled,
     },
     {
       name: 'sup',
-      active: state.superscriptActive,
+      active: state.superscriptActive || false,
       enabled: !state.superscriptDisabled,
     },
     {
       name: 'sub',
-      active: state.subscriptActive,
+      active: state.subscriptActive || false,
       enabled: !state.subscriptDisabled,
     },
   ];