+++ packages/editor/renderer/src/__tests__/visual-regression/__fixtures__/document-without-media.ts
@@ -184,11 +184,94 @@ export const document = {
         },
         {
           type: 'text',
-          text: ' red text',
+          text: ' red text ',
           marks: [
             {
               type: 'textColor',
-              attrs: { color: '#ff0000' },
+              attrs: {
+                color: '#ff0000',
+              },
+            },
+          ],
+        },
+        {
+          type: 'text',
+          text: 'strike-through',
+          marks: [
+            {
+              type: 'strike',
+            },
+            {
+              type: 'textColor',
+              attrs: {
+                color: '#ff0000',
+              },
+            },
+          ],
+        },
+        {
+          type: 'text',
+          text: ' ',
+          marks: [
+            {
+              type: 'textColor',
+              attrs: {
+                color: '#ff0000',
+              },
+            },
+          ],
+        },
+        {
+          type: 'text',
+          text: 'underline',
+          marks: [
+            {
+              type: 'underline',
+            },
+            {
+              type: 'textColor',
+              attrs: {
+                color: '#ff0000',
+              },
+            },
+          ],
+        },
+        {
+          type: 'text',
+          text: ' ',
+          marks: [
+            {
+              type: 'textColor',
+              attrs: {
+                color: '#ff0000',
+              },
+            },
+          ],
+        },
+        {
+          type: 'text',
+          text: 'mixed',
+          marks: [
+            {
+              type: 'strong',
+            },
+            {
+              type: 'strike',
+            },
+            {
+              type: 'subsup',
+              attrs: {
+                type: 'sup',
+              },
+            },
+            {
+              type: 'underline',
+            },
+            {
+              type: 'textColor',
+              attrs: {
+                color: '#ff0000',
+              },
             },
           ],
         },