+++ packages/editor/renderer/src/email/serializers.ts
@@ -20,6 +20,7 @@ import tableHeader from './nodes/table-header';
 import tableRow from './nodes/table-row';
 import text from './nodes/text';
 import unknownBlock from './nodes/unknown-block';
+import status from './nodes/status';
 
 import code from './marks/code';
 import em from './marks/em';
@@ -59,6 +60,7 @@ export const nodeSerializers: { [key: string]: NodeSerializer } = {
   taskList: renderNothing,
   text,
   unknownBlock,
+  status,
 };
 
 export const markSerializers: { [key: string]: MarkSerializer } = {