+++ packages/navigation/atlassian-switcher/src/index.ts
@@ -4,3 +4,6 @@ export {
 } from './components/confluence-switcher';
 export { SwitchToTooltipText } from './utils/messages';
 export { default } from './components/atlassian-switcher';
+export {
+  default as AtlassianSwitcherPrefetchTrigger,
+} from './components/prefetch-trigger';